Parity node tutorial: VPC; more CLI; corrections
Missing from the "How to Deploy a Parity Node" tutorial was any mention
of Virtual Private Cloud, which is a prerequiste for using
docker-machine.  This change gives CLI commands to do the VPC
configuration.

Also missing was a mention of installing Docker Machine, which is
separate from installing Docker itself.  This change adds a mention and
a link to installation instructions.

This change also adds command-line instructions for setting up the
Security Group, to avoid the labor of doing it manually.

Now corrected is an error in the `eval` command.

Finally, the ssh command used to access the machine is now given as
well.