Remove braces around environment variables
I wrote several new env var references in this doc without braces, and
now it's annoying me that things are not consistent.  Since braces are
superfluous anyways, I'm opting for the no-braces approach.  If we want
to have them for some reason, I'm happy to change it to yes-braces
everywhere.